refactor(memory): consolidate distill modules
