+++ src/lifecycle.ts
@@ -17,8 +17,7 @@ import { phasePrepare } from "./lifecycle-prepare";
 import { resolveModel } from "./lifecycle-resolve";
 import { createEmptyPromptBreakdownTotals } from "./lifecycle-usage";
 import type { MemoryCommitContext, MemoryCommitMetrics } from "./memory-contract";
-import { OBSERVER_PROMPT } from "./memory-distill-prompts";
-import { commitMemorySources } from "./memory-registry";
+import { commitMemorySources, OBSERVER_PROMPT } from "./memory-source-distill";
 import { createInMemoryTaskQueue } from "./task-queue";
 import { renderToolOutputPart } from "./tool-output-content";
 import { DISCOVERY_TOOL_SET, WRITE_TOOL_SET } from "./tool-registry";
