--- src/memory-distill-prompts.test.ts
@@ -1,9 +1,0 @@
-import { describe, test } from "bun:test";
-import { OBSERVER_PROMPT } from "./memory-distill-prompts";
-import { expectIntent } from "./test-utils";
-
-describe("memory distill prompts", () => {
-  test("observer prompt defines @observe scopes", () => {
-    expectIntent(OBSERVER_PROMPT, [["@observe project"], ["@observe user"], ["@observe session"]]);
-  });
-});
